componentize marquee bounding box
